+++ modules/core/src/main/scala/lucuma/odb/api/model/gmos/longslit/GmosLongslitMath.scala
@@ -82,11 +82,11 @@ trait GmosSiteSpecificLongslitMath extends GmosLongslitMath {
    *
    * @param dispersion - dispersion in nm/pix (see https://www.gemini.edu/sciops/instruments/gmos/spectroscopy-overview/gratings)
    */
-  def Δλ(dispersion: Quantity[Rational, NanometersPerPixel]): Quantity[Int, Nanometer] = {
+  def Δλ(dispersion: Quantity[Rational, NanometersPerPixel]): Quantity[BigDecimal, Nanometer] = {
     val d = dispersion.value.toDouble
     val g = gapSize.value.value
     val v = d * g * 2.0             // raw value, which we round to nearest 5 nm
-    ((v/5.0).round * 5.0).toInt.withUnit[Nanometer]
+    BigDecimal(((v/5.0).round * 5.0).toInt).withUnit[Nanometer]
   }
 
 }
